Rework of Plantform!
Changes the plantform to have more control over charts. We can now tell the chase the name of the value that the data is going to want to display.

Added an active not active button for the Plant
Added a get plant data button, that changes to refresh data afterwards. This is instead of using useEffect and gives more control to the user.